Rename main SDK package
