--- sdk/src/main/java/com/uid2/sdk/utils/TimeUtils.kt
@@ -1,17 +1,0 @@
-package com.uid2.sdk.utils
-
-/**
- * A class containing utility methods around the current time.
- */
-class TimeUtils {
-
-    /**
-     * Returns whether or not the given (epoch) time in milliseconds, is in the past.
-     */
-    fun hasExpired(expiryMs: Long) = expiryMs <= System.currentTimeMillis()
-
-    /**
-     * Returns the number of milliseconds difference between the given time and "now".
-     */
-    fun diffToNow(fromMs: Long) = fromMs - System.currentTimeMillis()
-}
